Conflicts:
	aiida/backends/djsite/db/subtests/query.py
	aiida/backends/querybuild/querybuilder_base.py
	aiida/backends/querybuild/querybuilder_django.py

+++ aiida/orm/implementation/sqlalchemy/node.py
@@ -153,14 +153,12 @@ class Node(AbstractNode):
         return q
 
     def _update_db_label_field(self, field_value):
-        from aiida.backends.sqlalchemy import session
         self.dbnode.label = field_value
         if not self._to_be_stored:
             self._dbnode.save(commit=False)
             self._increment_version_number_db()
 
     def _update_db_description_field(self, field_value):
-        from aiida.backends.sqlalchemy import session
         self.dbnode.description = field_value
         if not self._to_be_stored:
             self._dbnode.save(commit=False)
@@ -172,14 +170,19 @@ class Node(AbstractNode):
             self._add_dblink_from(src, label)
         except UniquenessError:
             # I have to replace the link; I do it within a transaction
-            self._remove_dblink_from(label)
-            self._add_dblink_from(src, label, link_type)
+            try:
+                self._remove_dblink_from(label)
+                self._add_dblink_from(src, label, link_type)
+                session.commit()
+            except:
+                session.rollback()
+                raise
 
     def _remove_dblink_from(self, label):
         from aiida.backends.sqlalchemy import session
-        link = self.dbnode.outputs.filter_by(label=label).first()
-        session.delete(link)
-        session.commit()
+        link = DbLink.query.filter_by(label=label).first()
+        if link is not None:
+            session.delete(link)
 
     def _add_dblink_from(self, src, label=None, link_type=LinkType.UNSPECIFIED):
         from aiida.backends.sqlalchemy import session
@@ -224,7 +227,7 @@ class Node(AbstractNode):
             safety_counter = 0
             while True:
                 safety_counter += 1
-                if safety_counter > 3:
+                if safety_counter > 100:
                     # Well, if you have more than 100 concurrent addings
                     # to the same node, you are clearly doing something wrong...
                     raise InternalError("Hey! We found more than 100 concurrent"
@@ -242,12 +245,11 @@ class Node(AbstractNode):
     def _do_create_link(self, src, label, link_type):
         from aiida.backends.sqlalchemy import session
         try:
-            link = DbLink(input_id=src.dbnode.id, output_id=self.dbnode.id,
-                          label=label, type=link_type.value)
-            session.add(link)
-            session.commit()
+            with session.begin_nested():
+                link = DbLink(input_id=src.dbnode.id, output_id=self.dbnode.id,
+                              label=label, type=link_type.value)
+                session.add(link)
         except SQLAlchemyError as e:
-            session.rollback()
             raise UniquenessError("There is already a link with the same "
                                   "name (raw message was {})"
                                   "".format(e))
@@ -382,9 +384,15 @@ class Node(AbstractNode):
         self.dbnode.del_extra(key)
 
     def extras(self):
+        if self.dbnode.extras is None:
+            return dict()
+
         return self.dbnode.extras
 
     def iterextras(self):
+        if self.dbnode.extras is None:
+            return dict().iteritems()
+
         return self.dbnode.extras.iteritems()
 
     def iterattrs(self):
